+++ NEWS.md
@@ -1,6 +1,7 @@
 # fmriqa 0.3.0
 * Improved method for calculating the object center of gravity for automated
 ROI placement.
+* Added method to calcuate the maximum background (MBG) signal.
 
 # fmriqa 0.2.0
 * Added an argument for the polynomial detrending order and changed the default