+++ NEWS.md
@@ -1,3 +1,6 @@
+# fmriqa 0.4.0
+* Edge detection is now used to improve object position detection.
+
 # fmriqa 0.3.0
 * Improved method for calculating the object center of gravity for automated
 ROI placement.